requirements.txt
edge-tts
openai
SpeechRecognition
pydub
python-dotenv
requests
pyaudio
